backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
PyMySQL>=1.1
openpyxl>=3.1
gunicorn>=21.2
django-cors-headers>=4.3
cryptography>=42.0.0
